% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{dt_to_ym}
\alias{dt_to_ym}
\title{Recode dates to a Month Year variable}
\usage{
dt_to_ym(dates, template = "{months} {years}")
}
\arguments{
\item{dates}{a date vector}
}
\description{
Recode dates to a Month Year variable
}
