% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{month_recode}
\alias{month_recode}
\title{Recode numbers to month labels}
\usage{
month_recode(x, abbr = T)
}
\arguments{
\item{x}{a numeric vector with numbers between 1 and 12}

\item{abbr}{abbreviate month labels?}
}
\description{
Recode numbers to month labels
}
