% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fct_recode.R
\name{prep_fct_recode}
\alias{prep_fct_recode}
\title{prep_fct_recode}
\usage{
prep_fct_recode(var, old_codes, new_codes)
}
\arguments{
\item{var}{a factor variable}

\item{old_codes}{character vector of old values to replace}

\item{new_codes}{character vector of new replacements (should have the same order as old values!)}
}
\value{

}
\description{
prep_fct_recode
}
\examples{
prep_fct_recode(diamonds$cut,c("Fair","Good"),c("Fair Diamond Quality","Pretty Good"))
}
