% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_freq_table.R
\name{prep_freq_table}
\alias{prep_freq_table}
\title{prep_freq_table
a function to calculate counts, proportions and conditional proportions based on grouping variables}
\usage{
prep_freq_table(data, ...)
}
\arguments{
\item{data}{a dataframe or tibble object}

\item{...}{arbitrary number of grouping variables}
}
\description{
prep_freq_table
a function to calculate counts, proportions and conditional proportions based on grouping variables
}
\examples{
library(ggplot2)
prep_freq_table(diamonds,cut,color)

}
