% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_pivot_table.R
\name{prep_pivot_table}
\alias{prep_pivot_table}
\title{Pivot Tables inspired by pandas pivot_tables function}
\usage{
prep_pivot_table(
  df,
  index = NULL,
  columns = NULL,
  outcomes,
  agg_funs = mean,
  fun_names = NULL
)
}
\arguments{
\item{df}{a dataframe or tibble}

\item{index}{"var", var, c("var") or c(var1,var2) to decribe the variables used to aggregated by row}

\item{columns}{"var", var, c("var") or c(var1,var2) to decribe the variables used to aggregated by column}

\item{outcomes}{"var", var, c("var") or c(var1,var2) to decribe the outcome variables}

\item{agg_funs}{"var", var, c("var") or c(var1,var2)}

\item{fun_names}{A string or character vector to prepend the column names for each function}
}
\description{
Pivot Tables inspired by pandas pivot_tables function
}
