% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepr-package.R
\docType{package}
\name{prepr-package}
\alias{prepr}
\alias{prepr-package}
\title{prepr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Gonrui/prepr}
  \item Report bugs at \url{https://github.com/Gonrui/prepr/issues}
}

}
\keyword{internal}
