% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace_many.R
\name{str_replace_many}
\alias{str_replace_many}
\title{Replace multiple strings in character vectors. Can be used for recoding!}
\usage{
str_replace_many(string, patterns, replacements)
}
\arguments{
\item{string}{A character vector}

\item{patterns}{A vector of strings or regular expressions to search for}

\item{replacements}{A vector of equal length to replace the found patterns}
}
\value{

}
\description{
Replace multiple strings in character vectors. Can be used for recoding!
}
