% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{to_year_month}
\alias{to_year_month}
\title{Recode dates to a Month Year variable}
\usage{
to_year_month(dates, template = "{months} {years}")
}
\arguments{
\item{dates}{a date vector}
}
\description{
Recode dates to a Month Year variable
}
